--- swift/Applist/Podfile
@@ -1,20 +1,0 @@
-source 'https://github.com/CocoaPods/Specs.git'
-
-platform :ios, '12.0'
-
-target 'Applist' do
-use_frameworks!
-
-    pod 'Alamofire'
-    pod 'Kingfisher' 
-
-    post_install do |installer|
-      installer.pods_project.targets.each do |target|
-          target.build_configurations.each do |config|
-              config.build_settings['SWIFT_VERSION'] = '5.4'
-          end
-      end
-    end
-end
-
-
